[feature #163652112] modals for critical actions and errors

+++ UI/admin/script.js
@@ -23,6 +23,8 @@ import createForm from '../helpers/createForm.js';
 import questions from '../helpers/questions.js';
 import notification from '../helpers/notification.js';
 import updateStats from '../helpers/updateStats.js';
+import confirmAct from '../helpers/confirmAct.js';
+import errorHandler from '../helpers/errorHandler.js';
 
 const tabSelector = document.getElementsByClassName('tab-selector');
 const profile = JSON.parse(window.sessionStorage.getItem('user'));
@@ -165,6 +167,10 @@ const getData = (inputClass) => {
   const inputFields = document.getElementsByClassName(inputClass);
   const obj = {};
   loop.call(inputFields, (input) => {
+    if (input.classList.contains('invalid')) {
+      errorHandler('check your data and try again');
+      return;
+    }
     const pointer = input.getAttribute('pointer');
     switch (pointer) {
       case 'tags':
@@ -201,14 +207,18 @@ const updateMeet = async (id) => {
     mergeTime(obj);
     const body = createForm(obj);
     try {
+      const decision = await confirmAct(
+        'Modify Meetup',
+        'Are you sure you want to continue',
+      );
+      if (!decision) return;
       const [data] = await fetchData.updateMeet(body, id);
       addMeet(data, true);
     } catch (error) {
-      console.log(error);
+      errorHandler(error);
       return;
     }
   }
-  // send patch request
   tabSelector[0].click();
   toggleOrganizer();
 };
@@ -236,12 +246,18 @@ const createMeet = async () => {
     tabSelector[0].click();
     clearInputs();
   } catch (error) {
-    console.log(error);
+    errorHandler(error);
+    errorHandler(error);
   }
 };
 
 const deleteMeet = async (id) => {
   try {
+    const decision = await confirmAct(
+      'Delete Meetup',
+      'Are you sure you want to continue',
+    );
+    if (!decision) return;
     await fetchData.deleteMeetup(id);
     loop.call(document.getElementsByClassName('meet-container'), (elem) => {
       if (elem.id === id.toString()) {
@@ -250,7 +266,7 @@ const deleteMeet = async (id) => {
     });
     tabSelector[0].click();
   } catch (error) {
-    console.log(error);
+    errorHandler(error);
   }
 };
 
@@ -323,7 +339,7 @@ const expandMeet = async (meetData) => {
     const profiles = await questions.get(meetData.id, box);
     profileControl(profiles, swith);
   } catch (error) {
-    console.log(error);
+    errorHandler(error);
   }
 };
 
@@ -332,14 +348,13 @@ const populateMeet = async () => {
     const meets = await fetchData.meetups();
     meets.forEach(meet => addMeet(meet));
   } catch (error) {
-    console.log(error);
+    errorHandler(error);
   }
 };
 
 const populateNotif = async () => {
   try {
     let data = await fetchData.notifications();
-    console.log(data);
     data = typeof data === 'string' ? [] : data;
     const elements = notifContainerCreator(document.querySelector('#notif-section'), data);
     elements.forEach((elem) => {
@@ -349,7 +364,7 @@ const populateNotif = async () => {
       notifContol(notifBtn);
     });
   } catch (error) {
-    console.log(error);
+    errorHandler(error);
   }
 };
 
@@ -396,9 +411,15 @@ const getResults = async (value) => {
     const [result] = await fetchData.search({ topic, tags });
     populateSearch(result);
   } catch (error) {
-    console.log(error);
+    errorHandler(error);
   }
 };
+
+document.getElementById('search-button').addEventListener('click', () => {
+  const input = document.getElementById('search-input').value;
+  if (!input) return;
+  getResults(input);
+});
 
 populateProfile(
   document.getElementsByClassName('profile-input'),
